+++ depthmapXcli/exportparser.h
@@ -42,7 +42,7 @@ public:
 
 public:
     ExportParser();
-    virtual void parse(int argc, char *argv[]);
+    virtual void parse(size_t argc, char *argv[]);
     virtual void run(const CommandLineParser &clp, IPerformanceSink& perfWriter) const;
 
     enum ExportMode{
