+++ depthmapXcli/importparser.h
@@ -43,7 +43,7 @@ public:
     }
 
 public:
-    virtual void parse(int argc, char *argv[]);
+    virtual void parse(size_t argc, char *argv[]);
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const;
 
     const std::vector<std::string> & getFilesToImport() const { return m_filesToImport; }
