+++ depthmapXcli/linkparser.h
@@ -60,7 +60,7 @@ public:
     }
 
 public:
-    virtual void parse(int argc, char *argv[]);
+    virtual void parse(size_t argc, char *argv[]);
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const;
 
     //link options
