+++ depthmapXcli/stepdepthparser.h
@@ -45,7 +45,7 @@ public:
         VISUAL
     };
 
-    virtual void parse(int argc, char** argv);
+    virtual void parse(size_t argc, char** argv);
 
     virtual void run(const CommandLineParser &clp, IPerformanceSink &perfWriter) const;
